+++ packages/studio-base/src/components/PanelExtensionAdapter/PanelExtensionAdapter.stories.tsx
@@ -9,8 +9,9 @@ import { toSec } from "@foxglove/rostime";
 import { PanelExtensionContext, ParameterValue, RenderState, Time } from "@foxglove/studio";
 import ErrorBoundary from "@foxglove/studio-base/components/ErrorBoundary";
 import MockPanelContextProvider from "@foxglove/studio-base/components/MockPanelContextProvider";
-import PanelExtensionAdapter from "@foxglove/studio-base/components/PanelExtensionAdapter";
 import PanelSetup from "@foxglove/studio-base/stories/PanelSetup";
+
+import PanelExtensionAdapter from "./PanelExtensionAdapter";
 
 export default {
   title: "PanelExtensionAdapter",
