+++ packages/studio-base/src/components/PanelExtensionAdapter/PanelExtensionAdapter.tsx
@@ -22,13 +22,11 @@ import { fromSec, toSec } from "@foxglove/rostime";
 import {
   AppSettingValue,
   ExtensionPanelRegistration,
-  MessageEvent,
   PanelExtensionContext,
   ParameterValue,
   RenderState,
   SettingsTree,
   Subscription,
-  Topic,
   VariableValue,
 } from "@foxglove/studio";
 import {
@@ -44,20 +42,17 @@ import {
   useHoverValue,
   useSetHoverValue,
 } from "@foxglove/studio-base/context/HoverValueContext";
-import useGlobalVariables, {
-  EMPTY_GLOBAL_VARIABLES,
-  GlobalVariables,
-} from "@foxglove/studio-base/hooks/useGlobalVariables";
+import useGlobalVariables from "@foxglove/studio-base/hooks/useGlobalVariables";
 import {
   AdvertiseOptions,
   PlayerCapabilities,
-  PlayerState,
   SubscribePayload,
 } from "@foxglove/studio-base/players/types";
 import { usePanelSettingsTreeUpdate } from "@foxglove/studio-base/providers/PanelSettingsEditorContextProvider";
-import { HoverValue } from "@foxglove/studio-base/types/hoverValue";
 import { PanelConfig, SaveConfig } from "@foxglove/studio-base/types/panels";
 import { assertNever } from "@foxglove/studio-base/util/assertNever";
+
+import { initRenderStateBuilder } from "./renderState";
 
 const log = Logger.getLogger(__filename);
 
@@ -72,183 +67,11 @@ type PanelExtensionAdapterProps = {
   help?: string;
 };
 
-const EmptyParameters = new Map<string, ParameterValue>();
-
-const EmptyTopics: readonly Topic[] = [];
-
 function selectContext(ctx: MessagePipelineContext) {
   return ctx;
 }
 
-type RenderFn = (renderState: Readonly<RenderState>, done: () => void) => void;
-
-function generateRender() {
-  let prevVariables: GlobalVariables = EMPTY_GLOBAL_VARIABLES;
-  let prevBlocks: unknown;
-  let prevSeekTime: number | undefined;
-  let prevSubscribedTopics: string[];
-
-  const prevRenderState: RenderState = {};
-
-  return function renderImpl(input: {
-    watchedFields: Set<string>;
-    playerState: PlayerState | undefined;
-    appSettings: Map<string, AppSettingValue> | undefined;
-    currentFrame: MessageEvent<unknown>[] | undefined;
-    colorScheme: RenderState["colorScheme"] | undefined;
-    globalVariables: GlobalVariables;
-    hoverValue: HoverValue | undefined;
-    subscribedTopics: string[];
-  }) {
-    const {
-      playerState,
-      watchedFields,
-      appSettings,
-      currentFrame,
-      colorScheme,
-      globalVariables,
-      hoverValue,
-      subscribedTopics,
-    } = input;
-
-    // If the player has loaded all the blocks, the blocks reference won't change so our message
-    // pipeline handler for allFrames won't create a new set of all frames for the newly
-    // subscribed topic. To ensure a new set of allFrames with the newly subscribed topic is
-    // created, we unset the blocks ref which will force re-creating allFrames.
-    if (subscribedTopics !== prevSubscribedTopics) {
-      prevBlocks = undefined;
-    }
-    prevSubscribedTopics = subscribedTopics;
-
-    // Should render indicates whether any fields of render state are updated
-    let shouldRender = false;
-
-    const activeData = playerState?.activeData;
-
-    // The render state starts with the previous render state and changes are applied as detected
-    const renderState: RenderState = prevRenderState;
-
-    if (watchedFields.has("didSeek")) {
-      const didSeek = prevSeekTime !== activeData?.lastSeekTime;
-      if (didSeek !== renderState.didSeek) {
-        renderState.didSeek = didSeek;
-        shouldRender = true;
-      }
-      prevSeekTime = activeData?.lastSeekTime;
-    }
-
-    if (watchedFields.has("currentFrame")) {
-      // If there are new frames we render
-      // If there are old frames we render (new frames either replace old or no new frames)
-      // Note: renderState.currentFrame.length !== currentFrame.length is wrong because it
-      // won't render when the number of messages is the same from old to new
-      if (renderState.currentFrame?.length !== 0 || currentFrame?.length !== 0) {
-        shouldRender = true;
-        renderState.currentFrame = currentFrame;
-      }
-    }
-
-    if (watchedFields.has("parameters")) {
-      const parameters = activeData?.parameters ?? EmptyParameters;
-      if (parameters !== renderState.parameters) {
-        shouldRender = true;
-        renderState.parameters = parameters;
-      }
-    }
-
-    if (watchedFields.has("variables")) {
-      if (globalVariables !== prevVariables) {
-        shouldRender = true;
-        prevVariables = globalVariables;
-        renderState.variables = new Map(Object.entries(globalVariables));
-      }
-    }
-
-    if (watchedFields.has("topics")) {
-      const newTopics = activeData?.topics ?? EmptyTopics;
-      if (newTopics !== prevRenderState.topics) {
-        shouldRender = true;
-        renderState.topics = newTopics;
-      }
-    }
-
-    if (watchedFields.has("allFrames")) {
-      // see comment for prevBlocksRef on why extended message store updates are gated this way
-      const newBlocks = playerState?.progress.messageCache?.blocks;
-      if (newBlocks && prevBlocks !== newBlocks) {
-        shouldRender = true;
-        const frames: MessageEvent<unknown>[] = (renderState.allFrames = []);
-        for (const block of newBlocks) {
-          if (!block) {
-            continue;
-          }
-
-          for (const messageEvents of Object.values(block.messagesByTopic)) {
-            for (const messageEvent of messageEvents) {
-              if (!subscribedTopics.includes(messageEvent.topic)) {
-                continue;
-              }
-              frames.push(messageEvent);
-            }
-          }
-        }
-      }
-      prevBlocks = newBlocks;
-    }
-
-    if (watchedFields.has("currentTime")) {
-      const currentTime = activeData?.currentTime;
-
-      if (currentTime != undefined && currentTime !== renderState.currentTime) {
-        shouldRender = true;
-        renderState.currentTime = currentTime;
-      } else {
-        if (renderState.currentTime != undefined) {
-          shouldRender = true;
-        }
-        renderState.currentTime = undefined;
-      }
-    }
-
-    if (watchedFields.has("previewTime")) {
-      const startTime = activeData?.startTime;
-
-      if (startTime != undefined && hoverValue != undefined) {
-        const stamp = toSec(startTime) + hoverValue.value;
-        if (stamp !== renderState.previewTime) {
-          shouldRender = true;
-        }
-        renderState.previewTime = stamp;
-      } else {
-        if (renderState.previewTime != undefined) {
-          shouldRender = true;
-        }
-        renderState.previewTime = undefined;
-      }
-    }
-
-    if (watchedFields.has("colorScheme")) {
-      if (colorScheme !== renderState.colorScheme) {
-        shouldRender = true;
-        renderState.colorScheme = colorScheme;
-      }
-    }
-
-    if (watchedFields.has("appSettings")) {
-      if (renderState.appSettings !== appSettings) {
-        shouldRender = true;
-        renderState.appSettings = appSettings;
-      }
-    }
-
-    if (!shouldRender) {
-      return undefined;
-    }
-
-    return renderState;
-  };
-}
-
+type RenderFn = NonNullable<PanelExtensionContext["onRender"]>;
 /**
  * PanelExtensionAdapter renders a panel extension via initPanel
  *
@@ -312,9 +135,9 @@ function PanelExtensionAdapter(props: PanelExtensionAdapterProps): JSX.Element {
   // This getter allows the extension context to remain stable through pipeline changes
   const getMessagePipelineContext = useMessagePipelineGetter();
 
-  // Generate render produces a function which computers the latest render state from a set of inputs
+  // initRenderStateBuilder render produces a function which computers the latest render state from a set of inputs
   // Spiritually its like a reducer
-  const [computeRenderState] = useState(() => generateRender());
+  const [buildRenderState, setBuildRenderState] = useState(() => initRenderStateBuilder());
 
   // Reset panel when config is cleared.
   useUpdateEffect(() => {
@@ -369,7 +192,7 @@ function PanelExtensionAdapter(props: PanelExtensionAdapterProps): JSX.Element {
       return;
     }
 
-    const renderState = computeRenderState({
+    const renderState = buildRenderState({
       watchedFields,
       globalVariables,
       hoverValue,
@@ -421,7 +244,7 @@ function PanelExtensionAdapter(props: PanelExtensionAdapterProps): JSX.Element {
     messageEvents,
     renderFn,
     colorScheme,
-    computeRenderState,
+    buildRenderState,
     globalVariables,
   ]);
 
@@ -607,6 +430,7 @@ function PanelExtensionAdapter(props: PanelExtensionAdapterProps): JSX.Element {
 
     // Reset local state when the panel element is mounted or changes
     setRenderFn(undefined);
+    setBuildRenderState(() => initRenderStateBuilder());
 
     const panelElement = document.createElement("div");
     panelElement.style.width = "100%";
